Fix: Update Railway CLI command syntax (#6)
This PR fixes the deployment issue by updating the Railway CLI command
syntax. The Railway CLI command format has changed, and the current
command in our workflow file was using incorrect syntax with --project=
and --service=. According to the Railway CLI documentation, the correct
format is to use the -s or --service flag followed by the service ID as
a separate value. This change will allow our deployments to work again
with the latest version of the Railway CLI.