Add support for API keys (#12)
As part of my effort to build a small display that shows the current
booking by the phone room (PR #11), I realized I would need a way for
the display to gain programmatic access to the bookings backend.

I tried using my personal Cookie on the device, which does work for
testing. However, I noticed that every time I would log in again on my
personal account, it would invalidate the token I had loaded on my
device. Therefore, I decided to build a more permanent solution, a
system to create and manage API keys.

This PR creates new end points to manage API keys, as follows:
```
POST /api/api-keys – Creates a new API key
GET /api/api-keys – List the user's API keys
DELETE /api/api-keys/:id – Delete the api key with the corresponding id
```

This PR also creates a new `api_keys` table in the database. This table
was added to the migration file, so it will be automatically created on
rollout.

This PR also modifies the authentication code, so that a user can
authenticate themselves with either their personal Cookie or an API key.

I've added several playwright tests to make sure that the the API keys
implementation works correctly, and that the previous Cookie-based
authentication still works as intended.

I expect the flow of a developer creating, using and deleting an API key
goes as follows:

```bash
$> curl -X POST http://localhost:3000/api/api-keys \
  -H "Cookie: auth_token=xxxxxxxxxxxxxxxxxxxxxxxxx" \
  -H "Content-Type: application/json" \
  -d '{"name": "My Integration"}'
{
    "id": 5,
    "key": "aeb89e73a96b030e6f4dd5b404a52b1e9d33cd7ba27bc122ccaeb36eda0007a5",
    "name": "My Integration",
    "prefix": "aeb89e73",
    "created_at": "2025-10-17T20:22:58.209Z"
}

$> curl http://localhost:3000/api/api-keys \
  -H "Cookie: auth_token=xxxxxxxxxxxxxxxxxxxxxxxxx"
[
    {
        "id": 5,
        "key_prefix": "aeb89e73",
        "name": "My Integration",
        "last_used_at": null,
        "created_at": "2025-10-17T20:22:58.209Z"
    },
]

$> curl http://localhost:3000/api/rooms \
  -H "Authorization: Bearer aeb89e73a96b030e6f4dd5b404a52b1e9d33cd7ba27bc122ccaeb36eda0007a5"
[
    {
        "id": 1,
        "name": "Green Phone Room",
        "description": "Small green phone booth for private calls",
        "capacity": 1,
        "created_at": "2025-10-12T21:42:44.510Z",
        "updated_at": "2025-10-12T21:42:44.510Z"
    },
    {
        "id": 2,
        "name": "Lovelace",
        "description": "Conference room named after Ada Lovelace",
        "capacity": 4,
        "created_at": "2025-10-12T21:42:44.510Z",
        "updated_at": "2025-10-12T21:42:44.510Z"
    }
]

$> curl -X DELETE http://localhost:3000/api/api-keys/5 \
  -H "Cookie: auth_token=xxxxxxxxxxxxxxxxxxxxxxxxx"
{
    "message":"API key deleted successfully"
}
```